+++ Lib/Data/Providers/InternetServices/Google.cs
@@ -32,7 +32,7 @@ namespace TrackConverter.Lib.Data.Providers.InternetServices
         /// </summary>
         /// <param name="cacheDirectory">папка с кэшем или null, если не надо использовать кэш</param>
         /// <param name="duration">время хранения кэша в часах. По умолчанию - неделя</param>
-        public Google(string cacheDirectory, int duration = 24 * 7) : base(cacheDirectory, duration) { }
+        public Google(string cacheDirectory, double duration = 24 * 7) : base(cacheDirectory, duration) { }
 
         /// <summary>
         /// временная папка для маршрутов
